Fix 64-bit -> 32-bit implicit conversions in libangle.
../../third_party/angle/src/libANGLE/renderer/gl/StateManagerGL.cpp(910,63):  warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long long') to 'GLuint' (aka 'unsigned int') [-Wshorten-64-to-32]
        const gl::ImageUnit &imageUnit = glState.getImageUnit(imageUnitIndex);
                                                 ~~~~~~~~~~~~ ^~~~~~~~~~~~~~
../../third_party/angle/src/libANGLE/renderer/gl/StateManagerGL.cpp(914,30):  warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long long') to 'GLuint' (aka 'unsigned int') [-Wshorten-64-to-32]
            bindImageTexture(imageUnitIndex, textureGL->getTextureID(), imageUnit.level,
            ~~~~~~~~~~~~~~~~ ^~~~~~~~~~~~~~
../../third_party/angle/src/libANGLE/renderer/gl/StateManagerGL.cpp(920,30):  warning: implicit conversion loses integer precision: 'size_t' (aka 'unsigned long long') to 'GLuint' (aka 'unsigned int') [-Wshorten-64-to-32]
            bindImageTexture(imageUnitIndex, 0, imageUnit.level, imageUnit.layered, imageUnit.layer,
            ~~~~~~~~~~~~~~~~ ^~~~~~~~~~~~~~

../../third_party/angle/src/tests/gl_tests/VertexAttributeTest.cpp(1080,66):  warning: implicit conversion loses integer precision: 'GLsizeiptr' (aka 'long long') to 'GLuint' (aka 'unsigned int') [-Wshorten-64-to-32]
        glVertexAttribFormat(mTestAttrib, 1, GL_FLOAT, GL_FALSE, inputRelativeOffset);
        ~~~~~~~~~~~~~~~~~~~~                                     ^~~~~~~~~~~~~~~~~~~

BUG=chromium:879657

+++ src/libANGLE/State.cpp
@@ -2859,7 +2859,7 @@ void State::onProgramExecutableChange(Program *program)
 }
 
 void State::setImageUnit(const Context *context,
-                         GLuint unit,
+                         size_t unit,
                          Texture *texture,
                          GLint level,
                          GLboolean layered,
@@ -2875,7 +2875,7 @@ void State::setImageUnit(const Context *context,
     mImageUnits[unit].format  = format;
 }
 
-const ImageUnit &State::getImageUnit(GLuint unit) const
+const ImageUnit &State::getImageUnit(size_t unit) const
 {
     return mImageUnits[unit];
 }
